# docs/api/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Lots API",
    "description": "Runs, approvals, audit, knowledge and administration. Authenticate with an OIDC access token or a personal API token (Authorization: Bearer lots_pat_...). See docs/api.md.",
    "version": "1"
  },
  "paths": {
...
    "/admin/voices/{userId}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "operationId": "LotsShellFeaturesVoiceRevokeOwnVoiceEndpoint",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
...
    "/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "operationId": "LotsShellFeaturesModelsListModelsEndpoint",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": [
          "V1"
        ],
        "operationId": "LotsShellFeaturesChannelsModelsListEndpoint",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelsDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelEndpointHealth"
            }
          },
          "aliases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AliasDto"
            }
          }
        }
      },
      "ModelEndpointHealth": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "up": {
            "type": "boolean"
          },
          "latencyMs": {
            "type": "integer",
            "format": "int64"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clearance": {
            "type": "string"
          }
        }
      },
      "AliasDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelTarget"
            }
          },
          "fastReasoningEffort": {
            "type": "string",
            "nullable": true
          },
          "usedBy": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ModelTarget": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "type": "http",
        "description": "Enter a JWT token to authorize the requests...",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}